### 2021-06-30 #3403 @Nikoleta-v3
:wave:  @MikeLydeamore &  @EduPH & @trashbirdecology & @wrathematics, I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS?

The submission I'd like you to review is titled: "Computer algebra in R". You can find more information at the top of this Github issue, and as a reminder the guidelines can be found at https://joss.readthedocs.io/en/latest/review_criteria.html

Please let me know if you're available :smile: Thank you!
