### 2021-07-01 #3438 @whedon
[REVIEW]: caracas: Computer algebra in R
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/r-cas/caracas" target ="_blank">https://github.com/r-cas/caracas</a>
**Version:** v1.1.1
