### 2021-07-05 r-cas/caracas#49 @Nikoleta-v3
JOSS review: typos 
Please resolve the following typos in the paper. I believe I've caught them all 😄

In **Statement of Need**

- [ ] `R` only support**s**
- [ ]  capapilities ->  capabilities
- [ ] Topicas -> Topics
- [ ] commerical -> commercial
- [ ] straight forward -> straightforward
- [ ]  which can, e.g.**,**

In **Acknowledgements**:

- [ ] pin pointing -> pinpointing (I think)
