### 2021-07-05 r-cas/caracas#50 @Nikoleta-v3
JOSS review: reference
> Several (commercial) systems are available for such tasks (and many more).

Any notable examples you could reference?
