### 2021-07-05 #3438 @Nikoleta-v3
@mikldk I opened some minor issues https://github.com/r-cas/caracas/issues/49 & https://github.com/r-cas/caracas/issues/50 (this is mostly nitpicking and spelling)
