+++ paper/paper.bib
@@ -74,3 +74,17 @@
  note = {\url{https://sdgs.un.org/}}
 }
 
+@misc{Mathematica,
+  author = {{Wolfram Research, Inc.}},
+  title = {Mathematica, {V}ersion 12.3.1},
+  url = {https://www.wolfram.com/mathematica},
+  note = {Champaign, IL, 2021},
+  year = {2021}
+}
+
+@misc{Maple,
+  author = {{Maplesoft, a division of Waterloo Maple Inc., Waterloo, Ontario}},
+  title = {Maple},
+  url = {https://www.maplesoft.com},
+  year = {2021}
+}
