### 2021-07-05 #3438 @mikldk
@Nikoleta-v3 : Thanks for catching these. I fixed [#49](https://github.com/r-cas/caracas/issues/49) in [379fb1](https://github.com/r-cas/caracas/commit/379fb1fa76a317edaa33681e65294c447aedba15) and [#50](https://github.com/r-cas/caracas/issues/50) in [ac6f07](https://github.com/r-cas/caracas/commit/ac6f07086525b94a63122b5fa3f3a83d64d3938e).
