### 2021-07-05 #3438 @Nikoleta-v3
Thank you @mikldk 👍🏻 Everything looks good to me.

`whedon` is suggesting the doi of this [book](https://link.springer.com/chapter/10.1007%2F978-3-211-73265-6_1) (which discusses both Maple and Mathematica) for the Maple reference. But your  current citations for both Maple and Mathematica are inline with the softwares' guidelines 👍🏻

At this point could you:
