### 2021-07-05 #3438 @mikldk
* The tagged release is `v.1.1.1` (as stated in the submission): <https://github.com/r-cas/caracas/releases/tag/v1.1.1>
* DOI: [10.5281/zenodo.5070324](https://doi.org/10.5281/zenodo.5070324)
* Zenodo: <https://zenodo.org/record/5070324>
