### 2021-07-05 #3438 @Nikoleta-v3
`joss-eics` team 👋🏻

Regarding the missing `doi`:

> Thank you @mikldk 👍🏻 Everything looks good to me.
>
> `whedon` is suggesting the doi of this [book](https://link.springer.com/chapter/10.1007%2F978-3-211-73265-6_1) (which discusses both Maple and Mathematica) for the Maple reference. But your current citations for both Maple and Mathematica are inline with the softwares' guidelines 👍🏻
>
> At this point could you:
>
> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [x]  Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission.
