### 2021-07-07 abf6518
expand_func and fix bug at sympy_func
