+++ man/sympy_func.Rd
@@ -29,6 +29,11 @@ if (has_sympy()) {
   expr <- x*y + x - 3 + 2*x^2 - z*x^2 + x^3
   expr
   expr \%>\% sympy_func("collect", x) 
+  
+  x <- symbol("x")
+  y <- gamma(x+3)
+  sympy_func(y, "expand_func")
+  expand_func(y)
 }
  
 }
