+++ NAMESPACE
@@ -49,6 +49,7 @@ export(eigenval)
 export(eigenvec)
 export(eval_to_symbol)
 export(expand)
+export(expand_func)
 export(expand_log)
 export(expand_trig)
 export(fraction_parts)
