### 2021-07-12 #3438 @Nikoleta-v3
I have removed the label just to be sure 😄 Nevertheless, `@openjournals/joss-eics` have been pinned by `whedon`. I am assuming it's only a matter of days until the paper is published
