### 2021-07-12 #3438 @kthyng
Hi @mikldk! A few comments on your paper before we wrap up:
* there is an errant space in the last sentence in the "summary" section.
* Please go through your references in detail to fix capitalization. You can preserve capitalization in bibtex using {} around characters, words, and phrases. For example, R, and Python are not always capitalized, but please check all entries.
