+++ paper/paper.bib
@@ -2,7 +2,7 @@
 	Address = {Vienna, Austria},
 	Author = {{R Core Team}},
 	Organization = {R Foundation for Statistical Computing},
-	Title = {R: A Language and Environment for Statistical Computing},
+	Title = {{R: A Language and Environment for Statistical Computing}},
 	Url = {https://www.R-project.org/},
 	Year = {2018},
 	Bdsk-Url-1 = {https://www.R-project.org/}}
@@ -16,13 +16,13 @@
   number = {42},
   pages = {1763},
   author = {Mikkel Meyer Andersen and Søren Højsgaard},
-  title = {Ryacas: A computer algebra system in R},
+  title = {{Ryacas: A computer algebra system in R}},
   journal = {Journal of Open Source Software}
 }
 
 @TechReport{yacas,
   author = 	 {Ayal Pinkus and Serge Winnitzky and Grzegorz Mazur},
-  title = 	 {Yacas - Yet another computer algebra system},
+  title = 	 {{Yacas - Yet another computer algebra system}},
   year = 	 {2016},
   Url = {https://yacas.readthedocs.io/en/latest/},
 }
@@ -43,7 +43,7 @@
 } 
 
 @Manual{reticulate,
-  title = {reticulate: Interface to 'Python'},
+  title = {{reticulate: Interface to 'Python'}},
   author = {Kevin Ushey and JJ Allaire and Yuan Tang},
   year = {2020},
   note = {R package version 1.18},
@@ -51,7 +51,7 @@
 }
 
 @article{sympy,
- title = {SymPy: symbolic computing in Python},
+ title = {{SymPy: symbolic computing in Python}},
  author = {Meurer, Aaron and Smith, Christopher P. and Paprocki, Mateusz and \v{C}ert\'{i}k, Ond\v{r}ej and Kirpichev, Sergey B. and Rocklin, Matthew and Kumar, AMiT and Ivanov, Sergiu and Moore, Jason K. and Singh, Sartaj and Rathnayake, Thilina and Vig, Sean and Granger, Brian E. and Muller, Richard P. and Bonazzi, Francesco and Gupta, Harsh and Vats, Shivam and Johansson, Fredrik and Pedregosa, Fabian and Curry, Matthew J. and Terrel, Andy R. and Rou\v{c}ka, \v{S}t\v{e}p\'{a}n and Saboo, Ashutosh and Fernando, Isuru and Kulal, Sumith and Cimrman, Robert and Scopatz, Anthony},
  year = 2017,
  month = jan,
@@ -70,13 +70,13 @@
 @misc{UN17,
  author = {{United Nations General Assembly}},
  year = {2015},
- title = {Sustainable Development Goals},
+ title = {{Sustainable Development Goals}},
  note = {\url{https://sdgs.un.org/}}
 }
 
 @misc{Mathematica,
   author = {{Wolfram Research, Inc.}},
-  title = {Mathematica, {V}ersion 12.3.1},
+  title = {{Mathematica, Version 12.3.1}},
   url = {https://www.wolfram.com/mathematica},
   note = {Champaign, IL, 2021},
   year = {2021}
@@ -84,7 +84,7 @@
 
 @misc{Maple,
   author = {{Maplesoft, a division of Waterloo Maple Inc., Waterloo, Ontario}},
-  title = {Maple},
+  title = {{Maple}},
   url = {https://www.maplesoft.com},
   year = {2021}
 }
