### 2021-07-18 #3438 @mikldk
@kthyng Thanks. I have now fixed these in https://github.com/r-cas/caracas/commit/53ff25db23928a3de035b817c1b1f94b1f99a4a8 . As this is only in the paper I am not sure if I should make a new Zenodo release or just keep the "old" one (with reviewed software, but out-of-sync paper source)?
