### 2021-07-18 0bd3527
Fix error with sympy_func
