### 2021-07-18 #3438 @danielskatz
You can choose - we don't require the paper source be in the repo at all, so either way is ok.
