### 2021-07-19 #3438 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
